Use ISO639 from http://www-01.sil.org/iso639-3/iso-639-3.tab
Instead of some handcrafted python code, just fetch iso-639-3.tab from the web,
convert it to a Python pickle object and store it as data. This makes for simple
code (not mixing data and code) and a nice updated language list.

AudioSelection.py: Don't import the module until actually needed.